requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=2.0.0
pydantic>=2.7.0
python-dotenv>=1.0.1
dateparser>=1.2.0
pillow>=10.4.0
pywin32>=306; platform_system == "Windows"
